backend/requirements.txt
# FastAPI e dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Upload de imagens
cloudinary==1.36.0
Pillow==10.1.0

# CORS e middleware
fastapi-cors==0.0.6

# Cache (opcional)
redis==5.0.1
aioredis==2.0.1

# Utilitários
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
